+++ src/hooks/mutations/useSaveSearch.tsx
@@ -3,7 +3,7 @@
  */
 
 import { gql, useMutation } from '@apollo/client';
-import { QueryableSearchFilterSet } from '@lib/classes';
+import { SearchFilterSetParams } from '@lib/types';
 import { QUERY_SAVED_SEARCHES } from '@hooks/queries/useSavedSearches';
 
 const MUTATE_UPDATE_SAVED_SEARCH = gql`
@@ -32,7 +32,7 @@ const useSaveSearch = () => {
 	}: {
 		id?: number;
 		title: string;
-		filterSet: QueryableSearchFilterSet;
+		filterSet: SearchFilterSetParams;
 		userId: number;
 	}) => {
 		return mutation({